+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Please view this file on the master branch, on other branches it's out of date.
 
 v1.6.2 (unreleased)
+  - Story templates by topic was not checking ?topic in URL
   - Media Gallery - Improved handling of video / audio playback for lightbox album theme
   - Set Auto Translation block to disabled on new installs
   - Removed tab slider example from home page on new installs